refactor: move SegmentationMerger to services port
Improve modularity and adhere to Dependency Injection Principle
